--- a/.gitignore
+++ b/.gitignore
@@ -29,3 +29,7 @@ zig-out/
 zig-cache/
 
 *.dot
+
+.gitignore/
+build-clang/
+build-icx/