--- a/examples/gpt-j/main.cpp
+++ b/examples/gpt-j/main.cpp
@@ -202,7 +202,7 @@ bool gptj_model_load(const std::string & fname, gptj_model & model, gpt_vocab & 
             /*.mem_size   =*/ ctx_size,
             /*.mem_buffer =*/ NULL,
             /*.no_alloc   =*/ false,
-            /*.n_threads  =*/ GGML_DEFAULT_N_THREADS,
+            /*.n_threads  =*/ 1,
         };
 
         model.ctx = ggml_init(params);