--- a/examples/gpt-j/quantize.cpp
+++ b/examples/gpt-j/quantize.cpp
@@ -143,7 +143,7 @@ int main(int argc, char ** argv) {
 
     // needed to initialize f16 tables
     {
-        struct ggml_init_params params = { 0, NULL, false };
+        struct ggml_init_params params = { 0, NULL, false, GGML_DEFAULT_N_THREADS };
         struct ggml_context * ctx = ggml_init(params);
         ggml_free(ctx);
     }