--- a/examples/gpt-neox/quantize.cpp
+++ b/examples/gpt-neox/quantize.cpp
@@ -139,7 +139,7 @@ int main(int argc, char ** argv) {
 
     // needed to initialize f16 tables
     {
-        struct ggml_init_params params = { 0, NULL, false, GGML_DEFAULT_N_THREADS };
+        struct ggml_init_params params = { 0, NULL, false, 1 };
         struct ggml_context * ctx = ggml_init(params);
         ggml_free(ctx);
     }