--- a/tests/test-grad0.c
+++ b/tests/test-grad0.c
@@ -345,6 +345,7 @@ int main(int argc, const char ** argv) {
         .mem_size   = 128*1024*1024,
         .mem_buffer = NULL,
         .no_alloc   = false,
+        .n_threads = GGML_DEFAULT_N_THREADS,
     };
 
     int64_t ne[4];