… examples/mnist/main.cpp
… examples/mpt/main.cpp
… examples/mpt/quantize.cpp
… examples/replit/main.cpp
… examples/replit/quantize.cpp
… examples/starcoder/main.cpp
… examples/starcoder/quantize.cpp
… examples/starcoder/starcoder-mmap.cpp
… examples/whisper/quantize.cpp
… examples/whisper/whisper.cpp
… include/ggml/pthread-win32.h
… include/ggml/pthreads-win32.h
… tests/test-blas0.c
… tests/test-mul-mat0.c
… tests/test-mul-mat2.c
… tests/test-opt.c
… tests/test-pool.c
… tests/test-quantize-fns.cpp
… tests/test-quantize-perf.cpp
… tests/test0.c
… tests/test1.c
… tests/test2.c
… tests/test3.c