Inspired by [this](https://github.com/ggerganov/llama.cpp/issues/2134) discussion I tried to understand the work scheduling. The patch uses a plain `pthreads` emulation for [Windows](https://nachtimwald.com/2019/04/05/cross-platform-thread-wrapper/) and a thread pool implementation from the same [site](https://nachtimwald.com/2019/04/12/thread-pool-in-c/) with what I hope are corrections. 

The patch is only tested on Windows. ~~When testing I noticed that `test_opt` seems to need quite a bit more run time than earlier. Is this a regression or due to a bug fix?~~ (This was due to different test sizes.)

Here are the results of a quick check with 

```
./bin/release/starcoder --threads 12 -m models/gpt_bigcode-santacoder/ggml-q4_0.bin -p "def test(" --temp 0
```

Before:

```
main: mem per token =   330184 bytes
main:     load time =  1162.09 ms
main:   sample time =    11.67 ms
main:  predict time = 18299.45 ms / 90.59 ms per token
main:    total time = 19597.86 ms
```

After:

```
main: mem per token =   330184 bytes
main:     load time =  1168.33 ms
main:   sample time =    13.62 ms
main:  predict time = 11272.42 ms / 55.80 ms per token
main:    total time = 12591.08 ms
``` 

Here the numbers for 

```
./bin/release/starcoder --threads 4 -m models/gpt_bigcode-santacoder/ggml-q4_0.bin -p "def test(" --temp 0
```

Before:

```
main: mem per token =   330056 bytes
main:     load time =  1189.20 ms
main:   sample time =    11.53 ms
main:  predict time =  9613.42 ms / 47.59 ms per token
main:    total time = 10968.48 ms
```

After:

``` 
main: mem per token =   330056 bytes
main:     load time =  1203.09 ms
main:   sample time =    12.97 ms
main:  predict time =  9725.99 ms / 48.15 ms per token
main:    total time = 11100.05 ms
```

Run on a Core i7 with 6 cores built with clang-cl and without BLAS support.