@slaren 2023-07-20
Tagging @JohannesGaessler since I think he was also planning on doing something similar.