@goerch 2023-07-21
> * If code is taken verbatim from an external source I think we should document this.

Absolutely. Assorted associations:

* I'd love to use some kind of standard `pthreads` thread pool, but this is the `best` I found
* The presentation on the website is expository
* I had to apply fixes to make the code compile
* I had to apply fixes to make the code run
* The code looks to me a bit like a text book implementation

> * The code in question is licensed as MIT so there should be no legal issues.

Maybe we should try to contact the author what to do?

Thanks for looking into this!