@goerch 2023-07-21
There are two more questions I'd like to discuss:

- Can we improve error handling?
- How would you like the condition variable to signal: during locks held or after releasing?