@JohannesGaessler 2023-07-22
On my Linux system with a Ryzen 3700X, 3200 MHz dual channel memory, and an RTX 3090 this PR has worse performance than master:

| GPU      | -ngl | Test  | ms/t master | ms/t PR | Speedup |
|----------|------|-------|-------------|---------|---------|
| RTX 3090 |    0 | tg128 |      402.21 |  426.87 |    0.94 |
| RTX 3090 |   20 | tg128 |      224.68 |  238.41 |    0.94 |
| RTX 3090 |   40 | tg128 |       42.62 |   51.55 |    0.83 |