@slaren 2023-07-23
Have you tested with llama? If I understand correctly, this will cause the threads to wait on a mutex for each node in the graph. In normal conditions, a spin lock should have lower acquire overhead than a mutex (excluding edge cases such as using more threads than physically available). Since that overhead is paid for every node, the overhead will scale with the number of nodes in the graph.
For what it's worth, what I was planning to do for https://github.com/ggerganov/llama.cpp/pull/2239 was to keep the spin locks while evaluating a graph, while keeping the threads alive waiting on a mutex between calls to `ggml_graph_compute`, to avoid having to re-launch them on every graph.