@ggerganov 2024-10-16
With the introduction of `ggml_threadpool` (https://github.com/ggerganov/llama.cpp/pull/8672) this is now obsolete.